Struct anneml::node::node::Node

pub struct Node { /* fields omitted */ }

Implementations

impl Node

pub fn new(uuid: Uuid, node_type: NodeType) -> Self

A Node is a wrapper for a Mesh.

The Node can either be a leaf or vertex, which will determine the role of its child Mesh. A Node with no children is a leaf- and vertex if it has children. Nodes connect between themselves, which form the structure of a Network.

Example(s)

use uuid::Uuid;
use anneml::node::attribute::Attribute;
use anneml::node::node::{Node, NodeType};
// Create a Leaf Node
let attribute = Attribute::default();
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let leaf_node = Node::new(Uuid::new_v4(), node_type);
assert_eq!(true, leaf_node.is_leaf_node());

// Create a Vertex Node
let node_type = NodeType::Vertex(attribute.clone(),leaf_node, attribute.clone());
let vertex_node = Node::new(Uuid::new_v4(), node_type);
assert_eq!(false, vertex_node.is_leaf_node());

pub fn add_child_to_parent(
&mut self,
child: Node,
index: Option<usize>
) -> Result<(), &'static str>

Assign a child Node to a parent Vertex node.

The passed in child Node is consumed and assigned to the parent. The index Option determines if the node will be appended to an existing column of children (Some) or a New Column (None)

Errors

Depending on whether the Node is a Leaf or a Vertex, we will return one of two results:

If the Node is a Vertex:

And index is Some(index) where index is within bounds of the current number of columns in the child vector, we will add a child to that column and return ok(())

And index is None, we append a new column with the child inside.

If the Node is a Leaf: We receive an Err of “Unable to add child to leaf.”.

Panics

Panics if index is out of bounds.

Example(s)

use uuid::Uuid;
use anneml::node::attribute::Attribute;
use anneml::node::node::{Node, NodeType};
// Add a new column
// Create a Leaf Node
let attribute = Attribute::default();
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let leaf_node = Node::new(Uuid::new_v4(), node_type);
// Create a Vertex Node, assign existing Leaf to Vertex Node
let node_type = NodeType::Vertex(attribute.clone(),leaf_node, attribute.clone());
let mut vertex_node = Node::new(Uuid::new_v4(), node_type);

// Create a NEW Leaf Node
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let new_leaf_node = Node::new(Uuid::new_v4(), node_type);

// Assign new leaf Node to new vertex_node child column
assert_eq!(1, vertex_node.children().len());
let result = vertex_node.add_child_to_parent(new_leaf_node, None);
assert_eq!(Ok(()), result);
assert_eq!(2, vertex_node.children().len());

use uuid::Uuid;
use anneml::node::attribute::Attribute;
use anneml::node::node::{Node, NodeType};
// Append to an existing column
// Create a Leaf Node
let attribute = Attribute::default();
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let leaf_node = Node::new(Uuid::new_v4(), node_type);
// Create a Vertex Node, assign existing Leaf to Vertex Node
let node_type = NodeType::Vertex(attribute.clone(),leaf_node, attribute.clone());
let mut vertex_node = Node::new(Uuid::new_v4(), node_type);

// Create a NEW Leaf Node
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let new_leaf_node = Node::new(Uuid::new_v4(), node_type);

// Assign new leaf Node to new vertex_node child column
assert_eq!(1, vertex_node.children()[0].len());
let result = vertex_node.add_child_to_parent(new_leaf_node, Some(0));
assert_eq!(Ok(()), result);
assert_eq!(2, vertex_node.children()[0].len());

use uuid::Uuid;
use anneml::node::attribute::Attribute;
use anneml::node::node::{Node, NodeType};
// Append add to leaf node
// Create a Leaf Node
let attribute = Attribute::default();
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let mut leaf_node = Node::new(Uuid::new_v4(), node_type);

// Create a NEW Leaf Node
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let new_leaf_node = Node::new(Uuid::new_v4(), node_type);

// Assign new leaf Node to new vertex_node child column
assert_eq!(0, leaf_node.children().len());
let result = leaf_node.add_child_to_parent(new_leaf_node, Some(0));
assert_eq!(Err("Unable to add child to leaf."), result);
assert_eq!(0, leaf_node.children().len());

use uuid::Uuid;
use anneml::node::attribute::Attribute;
use anneml::node::node::{Node, NodeType};
// Panic
// Create a Leaf Node
let attribute = Attribute::default();
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let leaf_node = Node::new(Uuid::new_v4(), node_type);
// Create a Vertex Node, assign existing Leaf to Vertex Node
let node_type = NodeType::Vertex(attribute.clone(),leaf_node, attribute.clone());
let mut vertex_node = Node::new(Uuid::new_v4(), node_type);

// Create a NEW Leaf Node
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let new_leaf_node = Node::new(Uuid::new_v4(), node_type);

// Assign new leaf Node to new vertex_node child column
let result = vertex_node.add_child_to_parent(new_leaf_node, Some(1));
assert!(false);

pub fn index_into_node(
&self,
indices: &[(usize, usize)]
) -> Result<Arc<Mutex<Node>>, &'static str>

Recursively Indexes into Nodes.

Indices are the (X,Y) position of the Node to index into and that nodes subsequent children.

Errors

Depending on whether the Node is a Leaf or a Vertex, we will return one of two results:

If the Node is a Vertex:

And indices: &[(usize, usize)] do not exceed the depth of the network, a Result<Arc<Mutex> will be returned.

And indices: &[(usize, usize)] does exceed the depth of the network, we panic.

If the Node is a Leaf: We receive an Err of “Unable to index into leaf node.”.

Panics

Panics if recursive index is out of bounds.

Example(s)

use uuid::Uuid;
use anneml::node::attribute::Attribute;
use anneml::node::node::{Node, NodeType};
// Index into a Vertex Nodes' child
// Create a Leaf Node
let attribute = Attribute::default();
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let leaf_node = Node::new(Uuid::new_v4(), node_type);
// Create a Vertex Node, assign existing Leaf to Vertex Node
let node_type = NodeType::Vertex(attribute.clone(),leaf_node, attribute.clone());
let mut vertex_node = Node::new(Uuid::new_v4(), node_type);

// Get Leaf Node &reference
let leaf_node = vertex_node.index_into_node(&[(0,0)]);
assert_eq!(true, leaf_node.unwrap().lock().unwrap().is_leaf_node());

use uuid::Uuid;
use anneml::node::attribute::Attribute;
use anneml::node::node::{Node, NodeType};
// Index into a Leaf Node
// Create a Leaf Node
let attribute = Attribute::default();
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let leaf_node = Node::new(Uuid::new_v4(), node_type);

// Assign new leaf Node to new vertex_node child column
let result = leaf_node.index_into_node(&[(0,0)]);
assert!(result.is_err());

use uuid::Uuid;
use anneml::node::attribute::Attribute;
use anneml::node::node::{Node, NodeType};
// Index into a Vertex Nodes' child
// Create a Leaf Node
let attribute = Attribute::default();
let node_type = NodeType::Leaf(attribute.clone(), vec![(attribute.clone(),2),(attribute.clone(),3)]);
let leaf_node = Node::new(Uuid::new_v4(), node_type);
// Create a Vertex Node, assign existing Leaf to Vertex Node
let node_type = NodeType::Vertex(attribute.clone(),leaf_node, attribute.clone());
let mut vertex_node = Node::new(Uuid::new_v4(), node_type);

// Get Leaf Node &reference (should panic)
let leaf_node = vertex_node.index_into_node(&[(0,0),(0,0)]);
assert!(false);

pub fn is_leaf_node(&self) -> bool

Returns whether a Node is a Leaf.

A Node is a leaf if it does not have any children.

pub fn output(&self) -> Array<f64>

Grabs the output values from children nodes and join them in a single column.

pub fn children(&self) -> &Vec<Vec<Arc<Mutex<Node>>>>

Acquire reference to Node Children Vector.